Crusader (#79)
* Fix (Core/Gossip) Trial of the crusader

* Adding enum with gossip and npc text

* Obtaining the npc_text with an emulator method

* fix build

* Modifying door closing event

From a logical point of view, when using the gossip, the door should close. Actually, before this change, the door closes, but only when the encounter is about to start. What I did, was to change the position, the place where the closing event is triggered. When the player, that the gossip, the door automatically closes. Preventing, that the event is activated and that the player can escape through the door.

* Modifying door opening events

Just as we have modified the closing events to prevent the player from escaping at the beginning of the event, we must also control that the door opens in relation to the end of the different dialogues.

Initially, the variable gossipMenuID was initialized in the constructor, however, when using a character of the horde, in the last event, it did not show me any dialogue to begin the last battle. So I moved the initialization.

* Get texts from the database

* Avoiding events if another event is in progress

The idea is to prevent the gossip from being available until the previous event has ended. Therefore, I removed the lines where the npc's gossip was enabled ahead of time, and reordered it within the door opening event. In other words, the door and the npc are available at the same time. That way, we get the event to end, then the door opens and also enables the npc for the next dialog.

Co-authored-by: Pagani Walter <paganiwalter@gmail.com>